Bugfix + Restructure
^
